Remove JLD2 as direct dependency
And update Manifests
