+++ Project.toml
@@ -25,7 +25,6 @@ ImageFiltering = "6a3955dd-da59-5b1f-98d4-e7296123deb5"
 Insolation = "e98cc03f-d57e-4e3c-b70c-8d51efe9e0d8"
 Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
 IntervalSets = "8197267c-284f-5f27-9208-e0e47529a953"
-JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 Krylov = "ba0b0d4f-ebba-5204-a429-3ac8c609bfb7"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
@@ -65,7 +64,6 @@ ImageFiltering = "0.7"
 Insolation = "0.6"
 Interpolations = "0.14"
 IntervalSets = "0.5, 0.6, 0.7"
-JLD2 = "0.4"
 Krylov = "0.9"
 NCDatasets = "0.11, 0.12, 0.13"
 NVTX = "0.3"
