+++ src/utils/utilities.jl
@@ -4,7 +4,6 @@
 import ClimaComms
 import ClimaCore: Spaces, Topologies, Fields, Geometry
 import LinearAlgebra: norm_sqr
-import JLD2
 
 is_energy_var(symbol) = symbol in (:ρθ, :ρe_tot, :ρaθ, :ρae_tot)
 is_momentum_var(symbol) = symbol in (:uₕ, :ρuₕ, :u₃, :ρw)
@@ -176,28 +175,6 @@ function time_to_seconds(s::String)
         return parse(Float64, first(split(s, match))) * factor[match]
     end
     error("Uncaught case in computing time from given string.")
-end
-
-function export_scaling_file(sol, output_dir, walltime, comms_ctx, nprocs)
-    # replace sol.u on the root processor with the global sol.u
-    if ClimaComms.iamroot(comms_ctx)
-        Y = sol.u[1]
-        center_space = axes(Y.c)
-        horz_space = Spaces.horizontal_space(center_space)
-        horz_topology = horz_space.topology
-        Nq = Spaces.Quadratures.degrees_of_freedom(horz_space.quadrature_style)
-        nlocalelems = Topologies.nlocalelems(horz_topology)
-        ncols_per_process = nlocalelems * Nq * Nq
-        scaling_file =
-            joinpath(output_dir, "scaling_data_$(nprocs)_processes.jld2")
-        @info(
-            "Writing scaling data",
-            "walltime (seconds)" = walltime,
-            scaling_file
-        )
-        JLD2.jldsave(scaling_file; nprocs, ncols_per_process, walltime)
-    end
-    return nothing
 end
 
 function verify_callbacks(t)
